move font enum out of remacs-sys
